Updated notebook, included CSV dataset for testing
The notebook now includes an if statement that checks if the user has the needed "tidyverse" and "tidycensus" package installs it, and installs it. Also the libraries are loaded now, and the census API key is now included. The CSV dataset has the same variables and the restriction of Oregon.